requirements.txt
langchain
langgraph
langchain-core
langchain-openai
pydantic
python-dotenv
argparse
requests